Create coreclr delegate inside coreclr_initialize
