+++ src/coreclr/hosts/unixcoreruncommon/coreruncommon.cpp
@@ -34,7 +34,6 @@
 // If set to 1, server GC is enabled on startup. If 0, server GC is
 // disabled. Server GC is off by default.
 static const char* serverGcVar = "CORECLR_SERVER_GC";
-GetInfoForMethodDelegate getInfoForMethodDelegate = NULL;
 #if defined(__linux__)
 #define symlinkEntrypointExecutable "/proc/self/exe"
 #elif !defined(__APPLE__)
@@ -402,24 +401,6 @@ int ExecuteManagedAssembly(
             }
             else 
             {
-#ifdef FEATURE_GDBJIT                
-                coreclr_create_delegate_ptr CreateDelegate =
-                (coreclr_create_delegate_ptr)dlsym(coreclrLib,
-                                           "coreclr_create_delegate");
-
-                // st = CreateDelegate(hostHandle, domainId, "System.Diagnostics.Debug.SymbolReader",
-                //           "System.Diagnostics.Debug.SymbolReader.SymbolReader", "GetLineByILOffset",
-                //           (void **)&getLineByILOffsetDelegate);
-                st = CreateDelegate(hostHandle, domainId, "System.Diagnostics.Debug.SymbolReader",
-                          "System.Diagnostics.Debug.SymbolReader.SymbolReader", "GetInfoForMethod",
-                          (void **)&getInfoForMethodDelegate);
-
-                if (!SUCCEEDED(st))
-                {
-                    fprintf(stderr, "coreclr_create_delegate failed - status: 0x%08x\n", st);
-                    exitCode = -1;
-                }
-#endif // FEATURE_GDBJIT                
                 st = executeAssembly(
                         hostHandle,
                         domainId,
